update: auth service

+++ backend/src/main/java/sit/cp23ej2/services/BookService.java
@@ -78,7 +78,7 @@ public class BookService extends CommonController {
                     // System.out.println("File name" + resource.getFile().getName());
                     // System.out.println("Resource"+ resource.getURI().toString());
                     Path pathFile = fileStorageService.load(bookDTO);
-                    if(pathFile != null){
+                    if (pathFile != null) {
                         bookDTO.setFile(pathFile.toString());
                     }
                 } catch (Exception e) {
@@ -103,7 +103,7 @@ public class BookService extends CommonController {
         if (book != null) {
             BookDTO bookDTO = modelMapper.map(book, BookDTO.class);
             try {
-    
+
                 Path pathFile = fileStorageService.load(bookDTO);
                 bookDTO.setFile(pathFile.toString());
             } catch (Exception e) {
@@ -140,36 +140,79 @@ public class BookService extends CommonController {
     }
 
     public DataResponse updateBook(UpdateBookDTO book, Integer bookId, MultipartFile file) {
+
         DataResponse response = new DataResponse();
-       
-        repository.updateBook(book.getBookName(), book.getAuthor(), book.getBookGenre(), book.getBookDetail(),
-                book.getBookRating(), bookId);
 
-        Book newDataBook = repository.getBookById(bookId);
-        BookDTO bookDTO = modelMapper.map(newDataBook, BookDTO.class);
+        Book dataBook = repository.getBookById(bookId);
 
-        if (file != null) {
-            Book dataBook = repository.getBookById(bookId);
-            fileStorageService.deleteFile(dataBook);
-            fileStorageService.store(file, book.getBookName(), book.getAuthor());
-        }
+        if(dataBook.getAuthor().equals(book.getAuthor()) && dataBook.getBookName().equals(book.getBookName())){
+            
+            repository.updateBook(book.getBookName(), book.getAuthor(), book.getBookGenre(), book.getBookDetail(),
+                    book.getBookRating(), bookId);
 
-        try {
-            if(book.getFilePath() != null){
-                Path pathFile = fileStorageService.load(bookDTO);
-                bookDTO.setFile(pathFile.toString());
-            }else{
-                fileStorageService.deleteFile(newDataBook);
+            Book newDataBook = repository.getBookById(bookId);
+            BookDTO bookDTO = modelMapper.map(newDataBook, BookDTO.class);
+
+            if (file != null) {
+                fileStorageService.deleteFile(dataBook);
+                fileStorageService.store(file, book.getBookName(), book.getAuthor());
             }
-        } catch (Exception e) {
-            e.printStackTrace();
+
+            try {
+                if (book.getStatus() != null) {
+                    Path pathFile = fileStorageService.load(bookDTO);
+                    bookDTO.setFile(pathFile.toString());
+                } else {
+                    fileStorageService.deleteFile(newDataBook);
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+
+            response.setResponse_code(200);
+            response.setResponse_status("OK");
+            response.setResponse_message("Book Updated");
+            response.setResponse_datetime(sdf3.format(new Timestamp(System.currentTimeMillis())));
+            response.setData(bookDTO);
+            return response;
+        }else{
+            Boolean existsByAuthorAndBookName = repository.existsByAuthorAndBookName(book.getAuthor(), book.getBookName());
+
+            if (!existsByAuthorAndBookName) {
+    
+                repository.updateBook(book.getBookName(), book.getAuthor(), book.getBookGenre(), book.getBookDetail(),
+                        book.getBookRating(), bookId);
+    
+                Book newDataBook = repository.getBookById(bookId);
+                BookDTO bookDTO = modelMapper.map(newDataBook, BookDTO.class);
+
+                if (file != null) {
+                    fileStorageService.deleteFile(dataBook);
+                    fileStorageService.store(file, book.getBookName(), book.getAuthor());
+                }
+    
+                try {
+                    if (book.getStatus() != null) {
+                        Path pathFile = fileStorageService.load(bookDTO);
+                        bookDTO.setFile(pathFile.toString());
+                    } else {
+                        fileStorageService.deleteFile(newDataBook);
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+    
+                response.setResponse_code(200);
+                response.setResponse_status("OK");
+                response.setResponse_message("Book Updated");
+                response.setResponse_datetime(sdf3.format(new Timestamp(System.currentTimeMillis())));
+                response.setData(bookDTO);
+                return response;
+            } else {
+                throw new HandleExceptionBadRequest("Book Already Exists");
+            }
         }
-        response.setResponse_code(200);
-        response.setResponse_status("OK");
-        response.setResponse_message("Book Updated");
-        response.setResponse_datetime(sdf3.format(new Timestamp(System.currentTimeMillis())));
-        response.setData(bookDTO);
-        return response;
+       
     }
 
     public DataResponse deleteBook(int bookId) {
@@ -177,11 +220,11 @@ public class BookService extends CommonController {
         try {
             Integer deleteStatus = repository.deleteBook(bookId);
             System.out.println("deleteBook: " + deleteStatus);
-            if(deleteStatus == 1){
+            if (deleteStatus == 1) {
                 Book dataBook = repository.getBookById(bookId);
                 fileStorageService.deleteFile(dataBook);
             }
-        }catch (Exception e) {
+        } catch (Exception e) {
             throw new HandleExceptionBadRequest("Book Can not delete because book have reviews.");
         }
         response.setResponse_code(200);
